requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
audioop-lts==0.2.2
cffi==2.0.0
cryptography==46.0.5
frozenlist==1.8.0
gmpy2==2.2.2
idna==3.11
multidict==6.7.1
nextcord==3.1.1
peewee==3.19.0
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
PyJWT==2.11.0
python-dotenv==1.2.1
python-ldap==3.4.5
setuptools==80.10.2
typing_extensions==4.15.0
Unidecode==1.4.0
yarl==1.22.0

fastecdsa